Use new xrefs from androguard
Using new get_xref functions avoids nested loops and decreases the
execution time by an order of magnitude on tested apps. However, since
the order of the generator is not preserved and the similarity score is
not stable for this ordering, a slight (<1%) change in similarity has
been observed when comparing the same apps before and after the changes.
Also removed returning the MethodAnalysis object that has been passed
to the affected functions.